Move all wire spec test to protocol folders
